remove v3rep and add into existing v3
