+++ iconrpcserver/default_conf/icon_rpcserver_constant.py
@@ -74,4 +74,3 @@ ICON_RPC_SERVER_LOG_TAG = 'IconService'
 DISPATCH_V2_TAG = f'{ICON_RPC_SERVER_LOG_TAG}_dispatch_v2'
 DISPATCH_V3_TAG = f'{ICON_RPC_SERVER_LOG_TAG}_dispatch_v3'
 DISPATCH_V3D_TAG = f'{ICON_RPC_SERVER_LOG_TAG}_dispatch_v3d'
-DISPATCH_V3REP_TAG = f'{ICON_RPC_SERVER_LOG_TAG}_dispatch_v3rep'
