+++ iconrpcserver/dispatcher/v3/version3.py
@@ -291,6 +291,28 @@ class Version3Dispatcher:
         return response_to_json_query(response)
 
     @staticmethod
+    @methods.add
+    async def rep_getList(**kwargs):
+        channel = kwargs['context']['channel']
+        channel_infos = await StubCollection().peer_stub.async_task().get_channel_infos()
+        channel_infos_by_channel = channel_infos.get(channel)
+
+        rep_list = [{'id': peer['id']} for peer in channel_infos_by_channel['peers']]
+        Logger.debug(f'rep list: {rep_list}')
+
+        start_term_height = '0x0'
+        end_term_height = '0x0'
+        rep_hash = None
+        # term_height, rep_root_hash should be updated after IISS is implemented.
+        response = {
+            'startTermHeight': start_term_height,
+            'endTermHeight': end_term_height,
+            'repHash': rep_hash,
+            'rep': rep_list
+        }
+        return response_to_json_query(response)
+
+    @staticmethod
     def _hash_convert(key: Optional[str], response: Any):
         if key is None:
             result = response
