+++ iconrpcserver/server/rest_server.py
@@ -30,7 +30,6 @@ from iconrpcserver.dispatcher.default import NodeDispatcher
 from iconrpcserver.dispatcher.v2 import Version2Dispatcher
 from iconrpcserver.dispatcher.v3 import Version3Dispatcher
 from iconrpcserver.dispatcher.v3d import Version3DebugDispatcher
-from iconrpcserver.dispatcher.v3rep import Version3RepDispatcher
 from ..utils.message_queue.stub_collection import StubCollection
 from sanic_cors import CORS
 from iconcommons.logger import Logger
@@ -93,9 +92,6 @@ class ServerComponents(metaclass=SingletonMetaClass):
 
         self.__app.add_route(Version3DebugDispatcher.dispatch, '/api/debug/v3/<channel_name>', methods=['POST'])
         self.__app.add_route(Version3DebugDispatcher.dispatch, '/api/debug/v3/', methods=['POST'])
-
-        self.__app.add_route(Version3RepDispatcher.dispatch, '/api/rep/v3', methods=['POST'])
-        self.__app.add_route(Version3RepDispatcher.dispatch, '/api/rep/v3/<channel_name>', methods=['POST'])
 
         self.__app.add_route(Disable.as_view(), '/api/v1', methods=['POST', 'GET'])
         self.__app.add_route(Status.as_view(), '/api/v1/status/peer')
